+++ gFunctionDatabase/Management/__init__.py
@@ -1,0 +1,2 @@
+from . import data_definition
+from . import retrieval